Split request and response range handling
